ActorsControlled modified with CustomBaseController
